Navigation folder and its content moved to components folder in common
