+++ src/common/components/StyledProfileIcon/index.js
@@ -1,4 +1,4 @@
-import { ReactComponent as ProfileIcon } from "../../../Profile.svg";
+import { ReactComponent as ProfileIcon } from "./Profile.svg";
 import styled from "styled-components";
 
 export const StyledProfileIcon = styled(ProfileIcon)`
